+++ src/client/Engine.ts
@@ -33,7 +33,7 @@ export default class Engine {
     private keysPressed: Set<string> = new Set<string>();
 
     private currentScene: string = 'main';
-    private sceneMap: SceneMap = new Map();
+    public sceneMap: SceneMap = new Map();
 
     private variableMap: Map<string, unknown> = new Map();