+++ src/client/sprites/Rectangle.ts
@@ -1,3 +1,4 @@
+import Engine from '../Engine.ts';
 import Sprite, { type BoundingBox, type SpriteOptions } from '../Sprite.ts';
 import { ctx, canvas, penCtx } from '../canvas.ts';
 
@@ -20,11 +21,24 @@ export default class Rectangle extends Sprite {
     public outlineWidth: number;
 
     public getBoundingBox(): BoundingBox {
+
+        const engine = Engine.init();
+
+        const w = this.width;
+        const h = this.height;
+        const dir = this.dir;
+
+        const cos = Math.abs(engine.cos(dir));
+        const sin = Math.abs(engine.sin(dir));
+
+        const bboxWidth = w * cos + h * sin;
+        const bboxHeight = w * sin + h * cos;
+
         return {
             x: this.x,
             y: this.y,
-            width: Math.hypot(this.width, this.height),
-            height: Math.hypot(this.width, this.height)
+            width: bboxWidth,
+            height: bboxHeight
         };
     }