+++ src/client/sprites/physics/RigidBodyOptions.ts
@@ -1,0 +1,8 @@
+export interface RigidBodyOptions {
+    gravity: number;
+    drag: number;
+    bounceLoss: number;
+    vX: number;
+    vY: number;
+    update: () => void;
+}